use common axes code
