+++ packages/sanddance/src/specs/stacks/axes.ts
@@ -1,25 +1,27 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 import { Axis } from 'vega-typings';
+import { partialAxes } from '../axes';
 import { SpecColumns, SpecViewOptions } from '../types';
 
 export default function (specViewOptions: SpecViewOptions, columns: SpecColumns) {
+    const pa = partialAxes(specViewOptions);
     const axes: Axis[] = [
         {
             "scale": "xband",
+            "title": columns.x.name,
+            "bandPosition": 0.5,
             "grid": true,
-            "domain": false,
-            "orient": "bottom",
-            "tickCount": 5,
-            "title": columns.x.name
+            "labelFlush": true,
+            ...pa.bottom as Axis
         },
         {
             "scale": "yband",
+            "title": columns.y.name,
+            "bandPosition": columns.y.quantitative ? 0 : 0.5,
             "grid": true,
-            "domain": false,
-            "orient": "left",
-            "titlePadding": 5,
-            "title": columns.y.name
+            "labelFlush": true,
+            ...pa.left as Axis
         }
     ];
     return axes;
